# pkg/graph/api/http/docs/swagger.yaml
basePath: /api
definitions:
...
  api.Graph:
    properties:
      attributes:
        additionalProperties: true
        description: Attrs are graph attributes.
        type: object
      edges:
        description: Edge is the edge count.
        type: integer
      label:
        description: Label is graph label.
        type: string
      nodes:
        description: Nodes is the node count.
        type: integer
      type:
        description: Type is a type of graph.
        type: string
      uid:
        description: UID is graph UUID.
        type: string
    type: object
  api.GraphUpdate:
    properties:
      attributes:
        additionalProperties: true
        type: object
      label:
        type: string
    type: object
...
  http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  http.GraphsResponse:
    properties:
      graphs:
        items:
          $ref: '#/definitions/api.Graph'
        type: array
      "n":
        type: integer
    type: object
...
info:
  contact:
    email: foo@bar.com
    name: API Support
  description: This is an API for managing Graphs.
  termsOfService: http://swagger.io/terms/
  title: Graphs API
  version: "1.0"
paths:
  /v1/graphs:
    get:
      description: Get all available graphs.
      parameters:
      - description: Result offset
        in: query
        name: offset
        type: integer
      - description: Result limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.GraphsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get all graphs
      tags:
      - graphs
    post:
      consumes:
      - application/json
      description: Create new graph.
      parameters:
      - description: Create a graph
        in: body
        name: graph
        required: true
        schema:
          $ref: '#/definitions/api.Graph'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Graph'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Create new graph.
      tags:
      - graphs
...
  /v1/graphs/{uid}:
    delete:
      description: Delete graph with the given UID.
      parameters:
      - description: Graph UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Graph was deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Delete graph by UID.
      tags:
      - graphs
    get:
      description: Get graph returns graph with the given UID.
      parameters:
      - description: Graph UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Graph'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get graph by UID.
      tags:
      - graphs
    patch:
      consumes:
      - application/json
      description: Update existing graph.
      parameters:
      - description: Graph UID
        in: path
        name: uid
        required: true
        type: string
      - description: Update a graph
        in: body
        name: graph
        required: true
        schema:
          $ref: '#/definitions/api.GraphUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Graph'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Update graph
      tags:
      - graphs
swagger: "2.0"